Supply_Chain_Optimization_Project/requirements.txt
pymysql
pandas
numpy
sqlalchemy
python-dotenv

# from sqlalchemy import create_engine
# import pandas as pd
# import pymysql
# import plotly.graph_objects as go
# from lightgbm import LGBMRegressor
# from sklearn.model_selection import train_test_split
# from sklearn.metrics import mean_squared_error